added comments to functions started fixing header file to only include what was needed to c file
