+++ src/entities/player_setup.c
@@ -10,7 +10,8 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cubed.h"
+#include "map.h"
+#include "state.h"
 
 void	player_rot_setup(char rot, t_player *player)
 {
@@ -62,6 +63,9 @@ void	player_pos_setup(t_list *curr, t_player *player)
 	}
 }
 
+/*
+ * the function is called within the map_setup.c
+ */
 t_player	player_setup(t_list *curr)
 {
 	t_player	player;
