+++ src/map_parser/map_checker_element.c
@@ -10,8 +10,13 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cubed.h"
+#include "map.h"
 
+/*
+ * checks the colors are in range and dont contain
+ * anything other then numbers. also checks the commer
+ * count to make sure the correct number of colors is passed
+ */
 int	map_check_element_colors(char *str)
 {
 	int		i;
@@ -27,7 +32,7 @@ int	map_check_element_colors(char *str)
 	while (rgb[i])
 	{
 		if (ft_atoi(rgb[i]) > 255 || ft_atoi(rgb[i]) < 0 \
-			|| is_line(rgb[i]))
+			|| is_empty_line(rgb[i]))
 			syntax_check = error_with("valueRange -", 1, 1);
 		j = 0;
 		while (rgb[i][j] && rgb[i][j] != '\n' && rgb[i][j] != ' ')
@@ -41,7 +46,12 @@ int	map_check_element_colors(char *str)
 	return (syntax_check);
 }
 
-/* str cant have spaces on end currently unsure if i should add that or not */
+/*
+ * checks the texture file can be opened and the file type is .xpm
+ *
+ * -str cant have spaces on end currently unsure if i should add that or not
+ * -because im doing a strcmp can change to strncmp 4
+ */
 int	map_check_element_texture(char *str)
 {
 	int	texture_file;
@@ -62,6 +72,10 @@ int	map_check_element_texture(char *str)
 	return (syntax_check);
 }
 
+/*
+ * checks elements are correct for the subject
+ * may need to add check if there is a double up of the same element
+ */
 int	map_check_elements(t_list *raw_map_files)
 {
 	t_list	*curr;
@@ -73,8 +87,6 @@ int	map_check_elements(t_list *raw_map_files)
 		str = (char *)curr->content;
 		if (map_starting_tile(str) == 1)
 			break ;
-	//	if (!str[1])
-	//		return (error_with("Empty Char", 1, 1));
 		if (str[2] && str[1] != ' ' && str[2] != ' ')
 			return (1);
 		if (str[1] && (str[0] == 'C' || str[0] == 'F') && str[1] == ' ')
