+++ src/map_parser/map_tile_processing.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cubed.h"
+#include "map.h"
 
 t_tiletype	get_tiletype(char c)
 {
@@ -22,6 +22,12 @@ t_tiletype	get_tiletype(char c)
 		return (EMPTY);
 }
 
+/*
+ * populates the map with the corrisponding tile there are only 3 types
+ * FLOOR/0 WALL/1 and everything else will be considered EMPTY/2
+ * could posisbly change to add more here but i would prefer to change them
+ * in modifiers so the map remains default as subject request
+ */
 int	map_tiles(t_map *map, char *content, int index)
 {
 	int	i;
@@ -41,6 +47,10 @@ int	map_tiles(t_map *map, char *content, int index)
 	return (i);
 }
 
+/*
+ * surrounds the whole map in empty spaces used to check boarder
+ * for error this is not included in gameplay map
+ */
 int	map_tiles_surround(t_map *map, char *content, int index)
 {
 	int	i;
