+++ src/map_parser/map_tools.c
@@ -10,8 +10,15 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cubed.h"
+#include "map.h"
 
+/*
+ * used to find the first instance of a wall you map be worried that
+ * it will get stopped in elements this should never happen as it 
+ * always checks the first character in content and each first character 
+ * in elements must consist of 1 or 2 uppercase letter, if it happen to 
+ * be a 1 this will get picked up in map_checker_element.c and return error
+ */
 int	map_starting_tile(char *content)
 {
 	int	i;
@@ -24,6 +31,10 @@ int	map_starting_tile(char *content)
 	return (0);
 }
 
+/*
+ * used to replace all taps with spaces to avoid error
+ * when parsing the map and modifiers
+ */
 void	replace_tabs(t_list *curr)
 {
 	int		i;
@@ -52,6 +63,10 @@ void	replace_tabs(t_list *curr)
 	}
 }
 
+/*
+ * a small fucntion used to print the map in terminal
+ * only useful for debugging not needed in release
+ */
 void	map_print(t_map *map)
 {
 	int	i;
@@ -66,6 +81,12 @@ void	map_print(t_map *map)
 	}
 }
 
+/*
+ * gets rid of(frees) the interger that hold the length of a
+ * link_list from the link_list there is also a small
+ * error check done here if the length of link list
+ * is less then or equal to 1 the file is empty
+ */
 void	remove_int_from_list(t_list **raw_map_file)
 {
 	t_list	*temp;
@@ -77,6 +98,10 @@ void	remove_int_from_list(t_list **raw_map_file)
 	free(temp);
 }
 
+/*
+ * gets rid of all the empty lines in the link_list to
+ * avoid any error while creating the map and modifiers
+ */
 void	remove_empty_lines(t_list **raw_map_file)
 {
 	t_list	*curr;
@@ -84,7 +109,7 @@ void	remove_empty_lines(t_list **raw_map_file)
 
 	remove_int_from_list(raw_map_file);
 	curr = *raw_map_file;
-	while (curr != NULL && is_line(curr->content) == 1)
+	while (curr != NULL && is_empty_line(curr->content) == 1)
 	{
 		temp = curr;
 		*raw_map_file = temp->next;
@@ -95,7 +120,7 @@ void	remove_empty_lines(t_list **raw_map_file)
 	{
 		if (map_starting_tile((char *)curr->content) == 1)
 			break ;
-		if (is_line(curr->next->content) == 1)
+		if (is_empty_line(curr->next->content) == 1)
 		{
 			temp = curr->next;
 			curr->next = temp->next;
