+++ src/modifiers/mod_func_cardinal_texture.c
@@ -10,8 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cubed.h"
-
+#include "map.h"
 
 void	mod_gen_no(char *content, int index, t_world *world, t_map *map)
 {
