+++ src/modifiers/modifier_setup.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cubed.h"
+#include "map.h"
 
 char *const			g_mapsymbols[] = {
 	"NO",
